+++ core/interfaces.go
@@ -262,6 +262,7 @@ type Storage interface {
 	DeleteQueueDataForMessagesWithContext(ctx context.Context, messagesIDs []string) error
 	DeleteQueueDataForRecipientsWithContext(ctx context.Context, recipientsIDs []string) error
 	DeleteQueueDataForUsers(ctx context.Context, orgID string, appID string, accountsIDs []string) error
+	DeleteFirebaseTokens(tokens []string) error
 }
 
 // Firebase is used to wrap all Firebase Messaging API functions