+++ core/app.go
@@ -19,7 +19,7 @@ import (
 	"notifications/driven/core"
 	"notifications/driven/mailer"
 
-	"github.com/rokwire/logging-library-go/v2/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
 )
 
 type storageListener struct {