+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [1.23.0] - 2024-17-04
+## [1.24.0] - 2025-09-007
+### Changed
+- Replace auth library and logging library with Building Block SDK [#196](https://github.com/rokwire/notifications-building-block/issues/196)
+
+## [1.23.0] - 2025-17-04
 ### Changed
 - Support Google Trust Services as CA [#193](https://github.com/rokwire/notifications-building-block/issues/193)